refactor(ui): extract APK build flow from HomeScreen into BuildApkScreen
Move the inline BuildApkDialog and ~1,080 lines of supporting build
composables/helpers out of HomeScreen.kt into a dedicated
BuildApkScreen.kt, reached via a new build_apk/{appId} navigation route
instead of a per-app dialog.

- HomeScreen: drop showBuildDialog/buildingApp state; add
  onBuildApk(Long) callback that navigates to the new route.
- AppNavigation: register BUILD_APK route; pass appId through the nav
  argument (replacing the prior load-full-WebApp-then-open coroutine).
- Lift the helpers shared with the share-APK failure-report path
  (BuildFailureReport, buildActionFailureReport,
  incrementalBuildModeLabel, buildDiagnosticLines, readBuildLogTail,
  BuildFailureReportDialog) to internal so HomeScreen keeps using them
  without forking copies.

No behavior change to the build/export pipeline; pure UI/architecture
extraction.

Closes #433
