+++ app/src/main/java/com/webtoapp/ui/screens/HomeScreen.kt
@@ -137,6 +137,7 @@ fun HomeScreen(
     onOpenPlayStore: () -> Unit = {},
     onExportAabForApp: (Long) -> Unit = {},
     onOpenFileManager: () -> Unit = {},
+    onBuildApk: (Long) -> Unit = {},
 ) {
     val apps by viewModel.filteredSummaries.collectAsStateWithLifecycle()
     val allWebApps by viewModel.webApps.collectAsStateWithLifecycle()
@@ -153,8 +154,6 @@ fun HomeScreen(
     var isSearchActive by remember { mutableStateOf(false) }
     var selectedApp by remember { mutableStateOf<WebAppSummary?>(null) }
     var showDeleteDialog by remember { mutableStateOf(false) }
-    var showBuildDialog by remember { mutableStateOf(false) }
-    var buildingApp by remember { mutableStateOf<WebApp?>(null) }
     var shareApkFailureReport by remember { mutableStateOf<BuildFailureReport?>(null) }
     var showFabMenu by remember { mutableStateOf(false) }
     var showBatchImportDialog by remember { mutableStateOf(false) }
@@ -626,11 +625,7 @@ fun HomeScreen(
                                 }
                             },
                             onBuildApk = {
-                                scope.launch {
-                                    val fullApp = viewModel.getWebApp(app.id) ?: return@launch
-                                    buildingApp = fullApp
-                                    showBuildDialog = true
-                                }
+                                onBuildApk(app.id)
                             },
                             onShareApk = {
                                 scope.launch {
@@ -868,29 +863,6 @@ fun HomeScreen(
                         }
                     }
             }
-    }
-
-    if (showBuildDialog && buildingApp != null) {
-        BuildApkDialog(
-            webApp = buildingApp!!,
-            onDismiss = {
-                showBuildDialog = false
-                buildingApp = null
-            },
-            onResult = { message ->
-                showBuildDialog = false
-                buildingApp = null
-                scope.launch {
-                    snackbarHostState.showSnackbar(message)
-                }
-            },
-            onExportAab = {
-                val appId = buildingApp?.id ?: return@BuildApkDialog
-                showBuildDialog = false
-                buildingApp = null
-                onExportAabForApp(appId)
-            }
-        )
     }
 
     shareApkFailureReport?.let { report ->
@@ -1552,1080 +1524,3 @@ private fun CreateActionTile(
     }
 }
 
-private data class BuildFailureReport(
-    val title: String,
-    val summary: String,
-    val details: String
-)
-
-private fun readBuildLogTail(path: String?, maxChars: Int = 20000): String {
-    return try {
-        path
-            ?.takeIf { it.isNotBlank() }
-            ?.let(::File)
-            ?.takeIf { it.exists() && it.isFile }
-            ?.readText()
-            ?.let { content ->
-                if (content.length <= maxChars) content else content.takeLast(maxChars)
-            }
-    } catch (e: Exception) {
-        AppLogger.e("HomeScreen", "读取 APK 构建日志失败", e)
-        Strings.readBuildLogFailed.format(e.message ?: "Unknown error")
-    } ?: "<build log unavailable>"
-}
-
-private fun buildActionFailureReport(
-    title: String,
-    stage: String,
-    webApp: WebApp,
-    summary: String,
-    logPath: String? = null,
-    throwable: Throwable? = null,
-    extraLines: List<String> = emptyList()
-): BuildFailureReport {
-    throwable?.let { AppLogger.e("HomeScreen", "$title failed at $stage", it) }
-    val buildLog = readBuildLogTail(logPath)
-
-    val details = buildString {
-        appendLine(title)
-        appendLine("stage: $stage")
-        appendLine("summary: $summary")
-        appendLine()
-        appendLine("project:")
-        appendLine("name=${webApp.name}")
-        appendLine("appType=${webApp.appType}")
-        appendLine("source=${webApp.url}")
-        if (extraLines.isNotEmpty()) {
-            appendLine()
-            appendLine("context:")
-            extraLines.forEach { appendLine(it) }
-        }
-        appendLine()
-        appendLine("log_path:")
-        appendLine(logPath ?: "<unavailable>")
-        appendLine()
-        appendLine("build_log:")
-        appendLine(buildLog)
-        if (throwable != null) {
-            appendLine()
-            appendLine("exception:")
-            appendLine(android.util.Log.getStackTraceString(throwable))
-        }
-        appendLine()
-        appendLine("recent_logs:")
-        append(AppLogger.getRecentLogTail())
-    }
-
-    return BuildFailureReport(
-        title = title,
-        summary = summary,
-        details = details
-    )
-}
-
-private fun buildBuildFailureReport(
-    webApp: WebApp,
-    error: BuildResult.Error
-): BuildFailureReport {
-    return buildActionFailureReport(
-        title = Strings.apkBuildFailed,
-        stage = error.diagnostic?.stage?.label ?: "apk_build",
-        webApp = webApp,
-        summary = error.message,
-        logPath = error.logPath,
-        extraLines = buildDiagnosticLines(error)
-    )
-}
-
-
-private fun incrementalBuildModeLabel(mode: String?): String {
-    return when (mode) {
-        "CONTENT_OVERLAY" -> Strings.buildModeContentOverlay
-        "REUSE_UNSIGNED" -> Strings.buildModeReuseUnsigned
-        "FULL" -> Strings.buildModeFull
-        else -> mode?.takeIf { it.isNotBlank() } ?: Strings.buildModeFull
-    }
-}
-
-private fun buildDiagnosticLines(error: BuildResult.Error): List<String> {
-    val diagnostic = error.diagnostic ?: return emptyList()
-    return buildList {
-        add("failureStage=${diagnostic.stage.name}")
-        add("failureCause=${diagnostic.cause.name}")
-        diagnostic.details.forEach { (key, value) ->
-            add("$key=$value")
-        }
-    }
-}
-
-@Composable
-private fun BuildFailureReportDialog(
-    report: BuildFailureReport,
-    onDismiss: () -> Unit
-) {
-    val clipboardManager = LocalClipboardManager.current
-
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = {
-            Column(verticalArrangement = Arrangement.spacedBy(4.dp)) {
-                Text(report.title)
-                Text(
-                    report.summary,
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-        },
-        text = {
-            EnhancedElevatedCard(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .heightIn(max = 420.dp),
-                containerColor = MaterialTheme.colorScheme.surfaceContainerHighest.copy(alpha = 0.55f)
-            ) {
-                Box(modifier = Modifier.fillMaxWidth()) {
-                    Text(
-                        text = report.details,
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(14.dp)
-                            .padding(bottom = 48.dp)
-                            .verticalScroll(rememberScrollState()),
-                        style = MaterialTheme.typography.bodySmall
-                    )
-
-                    FilledTonalButton(
-                        onClick = { clipboardManager.setText(AnnotatedString(report.details)) },
-                        modifier = Modifier
-                            .align(Alignment.BottomEnd)
-                            .padding(12.dp)
-                    ) {
-                        Icon(Icons.Outlined.ContentCopy, null, modifier = Modifier.size(16.dp))
-                        Spacer(modifier = Modifier.width(6.dp))
-                        Text(Strings.copy)
-                    }
-                }
-            }
-        },
-        confirmButton = {
-            TextButton(onClick = onDismiss) {
-                Text(Strings.close)
-            }
-        }
-    )
-}
-
-@Composable
-fun BuildApkDialog(
-    webApp: WebApp,
-    onDismiss: () -> Unit,
-    onResult: (String) -> Unit,
-    onExportAab: () -> Unit = {}
-) {
-    val context = LocalContext.current
-    val scope = rememberCoroutineScope()
-    val apkBuilder = remember { ApkBuilder(context) }
-    var showExportAabConfirm by remember { mutableStateOf(false) }
-
-    var isBuilding by remember(webApp.id) { mutableStateOf(false) }
-    var progress by remember(webApp.id) { mutableIntStateOf(0) }
-    var progressText by remember(webApp.id) { mutableStateOf(Strings.preparing) }
-    var analysisReport by remember(webApp.id) { mutableStateOf<com.webtoapp.core.apkbuilder.ApkAnalyzer.AnalysisReport?>(null) }
-    var buildFailureReport by remember(webApp.id) { mutableStateOf<BuildFailureReport?>(null) }
-    var preflightReport by remember(webApp.id) { mutableStateOf<ApkExportPreflightReport?>(null) }
-    var isEnsuringRuntime by remember(webApp.id) { mutableStateOf(false) }
-    var ensureRuntimeText by remember(webApp.id) { mutableStateOf<String?>(null) }
-    var forceFullRebuild by remember(webApp.id) { mutableStateOf(false) }
-    var lastBuildMode by remember(webApp.id) { mutableStateOf<String?>(null) }
-    var lastBuildReason by remember(webApp.id) { mutableStateOf<String?>(null) }
-    var cacheMessage by remember(webApp.id) { mutableStateOf<String?>(null) }
-
-    var encryptionConfig by remember(webApp.id) {
-        mutableStateOf(webApp.apkExportConfig?.encryptionConfig ?: com.webtoapp.data.model.ApkEncryptionConfig())
-    }
-
-    var isolationConfig by remember(webApp.id) {
-        mutableStateOf(resolveBuildIsolationDefault(webApp.apkExportConfig?.isolationConfig))
-    }
-
-    var backgroundRunEnabled by remember(webApp.id) {
-        mutableStateOf(webApp.apkExportConfig?.backgroundRunEnabled ?: false)
-    }
-    var backgroundRunConfig by remember(webApp.id) {
-        mutableStateOf(webApp.apkExportConfig?.backgroundRunConfig ?: com.webtoapp.data.model.BackgroundRunExportConfig())
-    }
-
-    var notificationEnabled by remember(webApp.id) {
-        mutableStateOf(webApp.apkExportConfig?.notificationEnabled ?: false)
-    }
-    var notificationConfig by remember(webApp.id) {
-        mutableStateOf(webApp.apkExportConfig?.notificationConfig ?: com.webtoapp.data.model.NotificationExportConfig())
-    }
-
-    var selectedEngineType by remember(webApp.id) {
-        mutableStateOf(webApp.apkExportConfig?.engineType ?: "SYSTEM_WEBVIEW")
-    }
-    val updatePackageName = webApp.apkExportConfig?.customPackageName
-        ?.takeIf { it.isNotBlank() && it.matches(com.webtoapp.util.AppConstants.PACKAGE_NAME_REGEX) }
-    val baseVersionCode = webApp.apkExportConfig?.customVersionCode ?: 1
-    val installedVersionCode = remember(updatePackageName) {
-        updatePackageName?.let { findInstalledVersionCode(context, it) }
-    }
-    val suggestedUpdateVersionCode = if (updatePackageName != null) {
-        (maxOf(baseVersionCode.toLong(), installedVersionCode ?: 0L) + 1L).coerceAtMost(Int.MAX_VALUE.toLong()).toInt()
-    } else {
-        baseVersionCode
-    }
-    val engineFileManager = remember { com.webtoapp.core.engine.download.EngineFileManager(context) }
-    val isGeckoDownloaded = remember(selectedEngineType) {
-        engineFileManager.isEngineDownloaded(com.webtoapp.core.engine.EngineType.GECKOVIEW)
-    }
-
-    fun currentBuildConfig(): WebApp {
-        return webApp.copy(
-            apkExportConfig = (webApp.apkExportConfig ?: com.webtoapp.data.model.ApkExportConfig()).copy(
-                encryptionConfig = encryptionConfig,
-                isolationConfig = isolationConfig,
-                backgroundRunEnabled = backgroundRunEnabled,
-                backgroundRunConfig = backgroundRunConfig,
-                notificationEnabled = notificationEnabled,
-                notificationConfig = notificationConfig,
-                engineType = selectedEngineType
-            ).let { exportConfig ->
-                if (updatePackageName != null && (exportConfig.customVersionCode ?: 1) < suggestedUpdateVersionCode) {
-                    exportConfig.copy(customVersionCode = suggestedUpdateVersionCode)
-                } else {
-                    exportConfig
-                }
-            }
-        ).withRuntimePermissionsSyncedFromFeatures()
-    }
-
-    fun launchBuild() {
-        if (isBuilding) return
-        val webAppWithConfig = currentBuildConfig()
-
-        isBuilding = true
-        buildFailureReport = null
-        analysisReport = null
-        lastBuildMode = null
-        lastBuildReason = null
-        cacheMessage = null
-        progress = 0
-        progressText = Strings.preparing
-        scope.launch {
-            progressText = when (webAppWithConfig.appType) {
-                com.webtoapp.data.model.AppType.PYTHON_APP -> Strings.preparingPythonEnv
-                com.webtoapp.data.model.AppType.NODEJS_APP -> Strings.preparingNodeEnv
-                com.webtoapp.data.model.AppType.PHP_APP,
-                com.webtoapp.data.model.AppType.WORDPRESS -> Strings.preparing
-                else -> Strings.preparing
-            }
-            val ensureOk = ExportRuntimeEnsure.ensure(
-                context,
-                webAppWithConfig.appType
-            )
-            if (!ensureOk) {
-                progressText = when (webAppWithConfig.appType) {
-                    com.webtoapp.data.model.AppType.PYTHON_APP -> Strings.pythonRuntimeDownloadFailed
-                    com.webtoapp.data.model.AppType.NODEJS_APP -> Strings.njsDownloadFailed
-                    com.webtoapp.data.model.AppType.PHP_APP,
-                    com.webtoapp.data.model.AppType.WORDPRESS -> Strings.wpDownloadFailed
-                    else -> Strings.preparing
-                }
-            }
-            val nextPreflight = ApkExportPreflight.check(context, webAppWithConfig)
-            preflightReport = nextPreflight
-            if (nextPreflight.hasErrors) {
-                isBuilding = false
-                return@launch
-            }
-
-            val result = apkBuilder.buildApk(
-                webApp = webAppWithConfig,
-                forceFullRebuild = forceFullRebuild
-            ) { p, t ->
-                progress = p
-                progressText = t
-            }
-            when (result) {
-                is BuildResult.Success -> {
-                    analysisReport = result.analysisReport
-                    lastBuildMode = result.buildMode
-                    lastBuildReason = result.buildReason.takeIf { it.isNotBlank() }
-                    progressText = Strings.buildModeUsed.replace(
-                        "%s",
-                        incrementalBuildModeLabel(result.buildMode)
-                    )
-                    isBuilding = false
-                }
-                is BuildResult.Error -> {
-                    buildFailureReport = buildBuildFailureReport(webAppWithConfig, result)
-                    isBuilding = false
-                }
-            }
-        }
-    }
-
-    LaunchedEffect(
-        webApp,
-        encryptionConfig,
-        isolationConfig,
-        backgroundRunEnabled,
-        backgroundRunConfig,
-        notificationEnabled,
-        notificationConfig,
-        selectedEngineType
-    ) {
-        val config = currentBuildConfig()
-        if (ExportRuntimeEnsure.needsEnsure(context, config.appType)) {
-            isEnsuringRuntime = true
-            ensureRuntimeText = when (config.appType) {
-                com.webtoapp.data.model.AppType.PYTHON_APP -> Strings.preparingPythonEnv
-                com.webtoapp.data.model.AppType.NODEJS_APP -> Strings.preparingNodeEnv
-                com.webtoapp.data.model.AppType.PHP_APP,
-                com.webtoapp.data.model.AppType.WORDPRESS -> Strings.preparing
-                else -> Strings.preparing
-            }
-            val ensureOk = ExportRuntimeEnsure.ensure(context, config.appType)
-            if (!ensureOk) {
-                ensureRuntimeText = when (config.appType) {
-                    com.webtoapp.data.model.AppType.PYTHON_APP -> Strings.pythonRuntimeDownloadFailed
-                    com.webtoapp.data.model.AppType.NODEJS_APP -> Strings.njsDownloadFailed
-                    com.webtoapp.data.model.AppType.PHP_APP,
-                    com.webtoapp.data.model.AppType.WORDPRESS -> Strings.wpDownloadFailed
-                    else -> Strings.preparing
-                }
-            } else {
-                ensureRuntimeText = null
-            }
-            isEnsuringRuntime = false
-        } else {
-            ensureRuntimeText = null
-            isEnsuringRuntime = false
-        }
-        preflightReport = ApkExportPreflight.check(context, config)
-    }
-
-    val dialogScrollState = rememberScrollState()
-    LaunchedEffect(analysisReport) {
-        if (analysisReport != null) {
-
-            withFrameNanos {}
-            dialogScrollState.animateScrollTo(dialogScrollState.maxValue)
-        }
-    }
-
-    AnimatedAlertDialog(
-        onDismissRequest = { if (!isBuilding) onDismiss() },
-        title = { Text(Strings.buildDialogTitle) },
-        text = {
-            Column(
-                verticalArrangement = Arrangement.spacedBy(12.dp),
-                modifier = Modifier.verticalScroll(dialogScrollState)
-            ) {
-
-                Row(verticalAlignment = Alignment.CenterVertically) {
-                    Box(
-                        modifier = Modifier
-                            .size(36.dp)
-                            .clip(RoundedCornerShape(com.webtoapp.ui.design.WtaRadius.IconPlate))
-                            .background(
-                                MaterialTheme.colorScheme.primary.copy(
-                                    alpha = com.webtoapp.ui.design.WtaAlpha.MutedContainer
-                                )
-                            ),
-                        contentAlignment = Alignment.Center
-                    ) {
-                        Icon(
-                            Icons.Outlined.Android,
-                            null,
-                            modifier = Modifier.size(20.dp),
-                            tint = MaterialTheme.colorScheme.onSurface
-                        )
-                    }
-                    Spacer(Modifier.width(12.dp))
-                    Column {
-                        Text(webApp.name, style = MaterialTheme.typography.titleSmall)
-                        Text(
-                            when (webApp.appType) {
-                                com.webtoapp.data.model.AppType.IMAGE -> {
-                                    webApp.mediaConfig?.mediaPath ?: webApp.url
-                                }
-                                com.webtoapp.data.model.AppType.VIDEO -> {
-                                    webApp.mediaConfig?.mediaPath ?: webApp.url
-                                }
-                                com.webtoapp.data.model.AppType.HTML -> {
-                                    webApp.htmlConfig?.entryFile?.takeIf { it.isNotBlank() } ?: "index.html"
-                                }
-                                else -> webApp.url
-                            },
-                            style = MaterialTheme.typography.bodySmall,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant,
-                            maxLines = 1,
-                            overflow = TextOverflow.Ellipsis
-                        )
-                    }
-                }
-
-                HorizontalDivider()
-
-                com.webtoapp.ui.components.EncryptionConfigCard(
-                    config = encryptionConfig,
-                    onConfigChange = { encryptionConfig = it }
-                )
-
-                com.webtoapp.ui.components.IsolationConfigCard(
-                    config = isolationConfig,
-                    onConfigChange = { isolationConfig = it }
-                )
-
-                com.webtoapp.ui.components.BackgroundRunConfigCard(
-                    enabled = backgroundRunEnabled,
-                    config = backgroundRunConfig,
-                    onEnabledChange = { backgroundRunEnabled = it },
-                    onConfigChange = { backgroundRunConfig = it }
-                )
-
-                com.webtoapp.ui.components.NotificationConfigCard(
-                    enabled = notificationEnabled,
-                    config = notificationConfig,
-                    onEnabledChange = { notificationEnabled = it },
-                    onConfigChange = { notificationConfig = it }
-                )
-
-                if (webApp.appType == com.webtoapp.data.model.AppType.WEB) {
-                    EngineSelectionCard(
-                        selectedEngine = selectedEngineType,
-                        isGeckoDownloaded = isGeckoDownloaded,
-                        onEngineSelected = { selectedEngineType = it }
-                    )
-                }
-
-                HorizontalDivider()
-
-                if (analysisReport == null && !isBuilding) {
-                    SettingsSwitch(
-                        title = Strings.forceFullRebuild,
-                        subtitle = Strings.forceFullRebuildDesc,
-                        checked = forceFullRebuild,
-                        onCheckedChange = { forceFullRebuild = it }
-                    )
-                    TextButton(
-                        onClick = {
-                            apkBuilder.clearIncrementalCache(currentBuildConfig())
-                            cacheMessage = Strings.incrementalCacheCleared
-                        },
-                        modifier = Modifier.fillMaxWidth()
-                    ) {
-                        Text(Strings.clearIncrementalCache)
-                    }
-                    cacheMessage?.let { msg ->
-                        Text(
-                            msg,
-                            style = MaterialTheme.typography.bodySmall,
-                            color = MaterialTheme.colorScheme.primary
-                        )
-                    }
-                }
-
-                if (analysisReport == null) {
-                    Text(
-                        Strings.buildApkForApp.replace("%s", webApp.name),
-                        style = MaterialTheme.typography.bodyMedium
-                    )
-
-                    Text(
-                        Strings.buildCompleteInstallHint,
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-
-                    if (updatePackageName != null) {
-                        Surface(
-                            shape = RoundedCornerShape(WtaRadius.Control),
-                            color = MaterialTheme.colorScheme.primaryContainer.copy(alpha = 0.55f)
-                        ) {
-                            Row(
-                                modifier = Modifier
-                                    .fillMaxWidth()
-                                    .padding(12.dp),
-                                horizontalArrangement = Arrangement.spacedBy(10.dp)
-                            ) {
-                                Icon(
-                                    Icons.Outlined.SystemUpdateAlt,
-                                    null,
-                                    modifier = Modifier.size(18.dp),
-                                    tint = MaterialTheme.colorScheme.primary
-                                )
-                                Text(
-                                    Strings.updateApkGuide.replace("%s", updatePackageName)
-                                        .replace("%d", suggestedUpdateVersionCode.toString()),
-                                    style = MaterialTheme.typography.bodySmall,
-                                    color = MaterialTheme.colorScheme.onPrimaryContainer
-                                )
-                            }
-                        }
-                    }
-                }
-
-                if (isEnsuringRuntime || ensureRuntimeText != null) {
-                    Surface(
-                        shape = RoundedCornerShape(WtaRadius.Control),
-                        color = if (isEnsuringRuntime) {
-                            MaterialTheme.colorScheme.primaryContainer.copy(alpha = 0.55f)
-                        } else {
-                            MaterialTheme.colorScheme.errorContainer.copy(alpha = 0.55f)
-                        }
-                    ) {
-                        Column(
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .padding(12.dp),
-                            verticalArrangement = Arrangement.spacedBy(8.dp)
-                        ) {
-                            Text(
-                                text = ensureRuntimeText ?: Strings.preparing,
-                                style = MaterialTheme.typography.bodySmall,
-                                color = if (isEnsuringRuntime) {
-                                    MaterialTheme.colorScheme.onPrimaryContainer
-                                } else {
-                                    MaterialTheme.colorScheme.onErrorContainer
-                                }
-                            )
-                            if (isEnsuringRuntime) {
-                                LinearProgressIndicator(modifier = Modifier.fillMaxWidth())
-                            }
-                        }
-                    }
-                }
-
-                if (!isEnsuringRuntime) {
-                    preflightReport?.let { report ->
-                        ApkExportPreflightPanel(report = report)
-                    }
-                }
-
-                if (isBuilding) {
-                    Spacer(Modifier.height(12.dp))
-
-                    val animatedProgress by animateFloatAsState(
-                        targetValue = progress / 100f,
-                        animationSpec = spring(
-                            dampingRatio = 0.8f,
-                            stiffness = Spring.StiffnessMediumLow
-                        ),
-                        label = "buildProgress"
-                    )
-
-                    val pulseAlpha by rememberInfiniteTransition(label = "buildPulse").animateFloat(
-                        initialValue = 0.6f,
-                        targetValue = 1f,
-                        animationSpec = infiniteRepeatable(
-                            animation = tween(1000),
-                            repeatMode = RepeatMode.Reverse
-                        ),
-                        label = "buildPulseAlpha"
-                    )
-                    val animPulse by animateFloatAsState(
-                        targetValue = pulseAlpha,
-                        animationSpec = tween(800),
-                        label = "pulseAlpha"
-                    )
-
-                    Row(
-                        verticalAlignment = Alignment.CenterVertically,
-                        horizontalArrangement = Arrangement.spacedBy(12.dp)
-                    ) {
-
-                        Box(contentAlignment = Alignment.Center) {
-                            CircularProgressIndicator(
-                                progress = { animatedProgress },
-                                modifier = Modifier.size(48.dp),
-                                strokeWidth = 4.dp,
-                                trackColor = MaterialTheme.colorScheme.primary.copy(alpha = 0.12f),
-                                color = MaterialTheme.colorScheme.primary.copy(alpha = animPulse)
-                            )
-                            Text(
-                                "${progress}%",
-                                style = MaterialTheme.typography.labelSmall,
-                                color = MaterialTheme.colorScheme.primary
-                            )
-                        }
-
-                        Column(modifier = Modifier.weight(1f)) {
-                            Text(
-                                progressText,
-                                style = MaterialTheme.typography.bodyMedium,
-                                color = MaterialTheme.colorScheme.onSurface
-                            )
-                            Spacer(Modifier.height(4.dp))
-                            LinearProgressIndicator(
-                                progress = { animatedProgress },
-                                modifier = Modifier.fillMaxWidth(),
-                                trackColor = MaterialTheme.colorScheme.primary.copy(alpha = 0.08f)
-                            )
-                        }
-                    }
-                }
-
-                analysisReport?.let { report ->
-                    HorizontalDivider()
-
-                    BuildSummaryCard(
-                        webApp = webApp,
-                        apkFile = report.apkFile,
-                        totalSizeFormatted = report.totalSizeFormatted,
-                        versionName = currentBuildConfig().apkExportConfig
-                            ?.customVersionName?.takeIf { it.isNotBlank() } ?: "1.0.0",
-                        versionCode = currentBuildConfig().apkExportConfig?.customVersionCode ?: 1,
-                        buildMode = lastBuildMode,
-                        buildReason = lastBuildReason
-                    )
-
-                    TextButton(
-                        onClick = {
-                            analysisReport = null
-                            lastBuildMode = null
-                            lastBuildReason = null
-                            buildFailureReport = null
-                            cacheMessage = null
-                        },
-                        modifier = Modifier.fillMaxWidth()
-                    ) {
-                        Text(Strings.buildAgain)
-                    }
-
-                    HorizontalDivider()
-
-                    Row(
-                        modifier = Modifier.fillMaxWidth(),
-                        horizontalArrangement = Arrangement.SpaceBetween,
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        Text(
-                            "APK Analysis",
-                            style = MaterialTheme.typography.titleSmall,
-                            color = MaterialTheme.colorScheme.primary
-                        )
-                        Text(
-                            report.totalSizeFormatted,
-                            style = MaterialTheme.typography.titleSmall,
-                            fontWeight = androidx.compose.ui.text.font.FontWeight.SemiBold
-                        )
-                    }
-
-                    Spacer(Modifier.height(4.dp))
-
-                    report.categories.forEach { cat ->
-                        val catColor = try {
-                            androidx.compose.ui.graphics.Color(android.graphics.Color.parseColor(cat.category.color))
-                        } catch (_: Exception) {
-                            MaterialTheme.colorScheme.primary
-                        }
-                            Row(
-                                modifier = Modifier.fillMaxWidth(),
-                                verticalAlignment = Alignment.CenterVertically
-                            ) {
-                                Box(
-                                    modifier = Modifier
-                                        .size(8.dp)
-                                        .background(catColor, RoundedCornerShape(WtaRadius.Button))
-                                )
-                                Spacer(Modifier.width(6.dp))
-                                Text(
-                                cat.category.displayName,
-                                style = MaterialTheme.typography.labelSmall,
-                                modifier = Modifier.weight(weight = 1f, fill = true)
-                            )
-                            Text(
-                                String.format("%.1f%%", cat.percentage),
-                                style = MaterialTheme.typography.labelSmall,
-                                color = MaterialTheme.colorScheme.onSurfaceVariant
-                            )
-                        }
-
-                        LinearProgressIndicator(
-                            progress = { (cat.percentage / 100f).coerceIn(0f, 1f) },
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .height(4.dp)
-                                .padding(start = 14.dp)
-                                .clip(RoundedCornerShape(WtaRadius.Button)),
-                            color = catColor,
-                            trackColor = MaterialTheme.colorScheme.surfaceVariant
-                        )
-                        Spacer(Modifier.height(2.dp))
-                    }
-                }
-            }
-        },
-        confirmButton = {
-            if (!isBuilding) {
-                val builtApk = analysisReport?.apkFile
-                Row(verticalAlignment = Alignment.CenterVertically) {
-                    TextButton(onClick = { showExportAabConfirm = true }) {
-                        Icon(
-                            Icons.Outlined.PlayCircleOutline,
-                            null,
-                            modifier = Modifier.size(18.dp)
-                        )
-                        Spacer(Modifier.width(6.dp))
-                        Text("AAB", maxLines = 1)
-                    }
-                    Spacer(Modifier.width(4.dp))
-                    TextButton(onClick = onDismiss) {
-                        Text(if (analysisReport != null) Strings.close else Strings.btnCancel, maxLines = 1)
-                    }
-                    Spacer(Modifier.width(8.dp))
-                    PremiumButton(
-                        onClick = {
-                            if (builtApk != null) {
-
-                                val installStarted = apkBuilder.installApk(builtApk)
-                                onResult(
-                                    if (installStarted) "正在启动安装..."
-                                    else "无法自动启动安装"
-                                )
-                            } else {
-                                launchBuild()
-                            }
-                        },
-                        enabled = !isEnsuringRuntime
-                    ) {
-                        val icon = if (builtApk != null) Icons.Outlined.GetApp else Icons.Outlined.Build
-                        Icon(icon, null, Modifier.size(18.dp))
-                        Spacer(Modifier.width(8.dp))
-                        Text(
-                            when {
-                                builtApk != null -> Strings.install
-                                buildFailureReport != null || preflightReport?.hasErrors == true -> Strings.btnRetry
-                                else -> Strings.btnStartBuild
-                            },
-                            maxLines = 1,
-                            overflow = TextOverflow.Ellipsis
-                        )
-                    }
-                }
-            } else {
-                CircularProgressIndicator(modifier = Modifier.size(24.dp))
-            }
-        }
-    )
-
-    if (showExportAabConfirm) {
-        androidx.compose.material3.AlertDialog(
-            onDismissRequest = { showExportAabConfirm = false },
-            icon = { Icon(Icons.Outlined.PlayCircleOutline, null) },
-            title = { Text(Strings.playStoreExportAabConfirmTitle) },
-            text = { Text(Strings.playStoreExportAabConfirmBody) },
-            confirmButton = {
-                androidx.compose.material3.TextButton(onClick = {
-                    showExportAabConfirm = false
-                    onExportAab()
-                }) { Text(Strings.confirm) }
-            },
-            dismissButton = {
-                androidx.compose.material3.TextButton(onClick = {
-                    showExportAabConfirm = false
-                }) { Text(Strings.btnCancel) }
-            }
-        )
-    }
-
-    buildFailureReport?.let { report ->
-        BuildFailureReportDialog(
-            report = report,
-            onDismiss = { buildFailureReport = null }
-        )
-    }
-}
-
-private fun resolveBuildIsolationDefault(
-    config: com.webtoapp.core.privacy.IsolationConfig?
-): com.webtoapp.core.privacy.IsolationConfig {
-    return config ?: com.webtoapp.core.privacy.IsolationConfig.DISABLED
-}
-
-private fun findInstalledVersionCode(context: android.content.Context, packageName: String): Long? {
-    return try {
-        val packageInfo = if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.TIRAMISU) {
-            context.packageManager.getPackageInfo(
-                packageName,
-                android.content.pm.PackageManager.PackageInfoFlags.of(0)
-            )
-        } else {
-            @Suppress("DEPRECATION")
-            context.packageManager.getPackageInfo(packageName, 0)
-        }
-        if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.P) {
-            packageInfo.longVersionCode
-        } else {
-            @Suppress("DEPRECATION")
-            packageInfo.versionCode.toLong()
-        }
-    } catch (_: Exception) {
-        null
-    }
-}
-
-@Composable
-private fun BuildSummaryCard(
-    webApp: WebApp,
-    apkFile: java.io.File,
-    totalSizeFormatted: String,
-    versionName: String,
-    versionCode: Int,
-    buildMode: String? = null,
-    buildReason: String? = null
-) {
-    val context = LocalContext.current
-    val clipboardManager = LocalClipboardManager.current
-
-    Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-        Text(
-            Strings.buildSummaryTitle,
-            style = MaterialTheme.typography.titleSmall,
-            color = MaterialTheme.colorScheme.primary
-        )
-
-        Row(verticalAlignment = Alignment.CenterVertically) {
-            val iconPath = webApp.iconPath
-            if (!iconPath.isNullOrBlank()) {
-                AsyncImage(
-                    model = iconPath,
-                    contentDescription = null,
-                    modifier = Modifier
-                        .size(40.dp)
-                        .clip(RoundedCornerShape(WtaRadius.IconPlate)),
-                    contentScale = ContentScale.Crop
-                )
-            } else {
-                Box(
-                    modifier = Modifier
-                        .size(40.dp)
-                        .clip(RoundedCornerShape(WtaRadius.IconPlate))
-                        .background(
-                            MaterialTheme.colorScheme.primary.copy(
-                                alpha = com.webtoapp.ui.design.WtaAlpha.MutedContainer
-                            )
-                        ),
-                    contentAlignment = Alignment.Center
-                ) {
-                    Icon(
-                        Icons.Outlined.Android, null,
-                        modifier = Modifier.size(22.dp),
-                        tint = MaterialTheme.colorScheme.onSurface
-                    )
-                }
-            }
-            Spacer(Modifier.width(12.dp))
-            Text(
-                webApp.name,
-                style = MaterialTheme.typography.titleSmall,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis
-            )
-        }
-
-        BuildSummaryRow(label = Strings.buildSummaryAppSize, value = totalSizeFormatted)
-        BuildSummaryRow(label = Strings.buildSummaryVersion, value = "$versionName ($versionCode)")
-        if (buildMode != null) {
-            BuildSummaryRow(
-                label = Strings.buildSummaryMode,
-                value = incrementalBuildModeLabel(buildMode)
-            )
-        }
-        if (!buildReason.isNullOrBlank()) {
-            BuildSummaryRow(
-                label = Strings.buildSummaryReason,
-                value = buildReason
-            )
-        }
-        BuildSummaryRow(label = Strings.buildSummaryJdk, value = "17")
-        BuildSummaryRow(label = Strings.buildSummarySignature, value = "v1+v2+v3")
-
-        Column(verticalArrangement = Arrangement.spacedBy(4.dp)) {
-            Text(
-                Strings.buildSummaryLocation,
-                style = MaterialTheme.typography.labelSmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-            Text(
-                apkFile.absolutePath,
-                style = MaterialTheme.typography.bodySmall,
-                color = MaterialTheme.colorScheme.onSurface,
-                maxLines = 2,
-                overflow = TextOverflow.Ellipsis
-            )
-            Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-                FilledTonalButton(
-                    onClick = {
-                        clipboardManager.setText(AnnotatedString(apkFile.absolutePath))
-                        android.widget.Toast.makeText(
-                            context, Strings.copiedToClipboard,
-                            android.widget.Toast.LENGTH_SHORT
-                        ).show()
-                    },
-                    contentPadding = androidx.compose.foundation.layout.PaddingValues(
-                        horizontal = 12.dp, vertical = 6.dp
-                    )
-                ) {
-                    Icon(Icons.Outlined.ContentCopy, null, Modifier.size(16.dp))
-                    Spacer(Modifier.width(6.dp))
-                    Text(Strings.buildSummaryCopyPath, style = MaterialTheme.typography.labelMedium)
-                }
-                FilledTonalButton(
-                    onClick = { openApkWithChooser(context, apkFile) },
-                    contentPadding = androidx.compose.foundation.layout.PaddingValues(
-                        horizontal = 12.dp, vertical = 6.dp
-                    )
-                ) {
-                    Icon(Icons.Outlined.Share, null, Modifier.size(16.dp))
-                    Spacer(Modifier.width(6.dp))
-                    Text(Strings.buildSummaryOpenWith, style = MaterialTheme.typography.labelMedium)
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun BuildSummaryRow(label: String, value: String) {
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Text(
-            label,
-            style = MaterialTheme.typography.labelSmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-        Text(
-            value,
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurface,
-            maxLines = 1,
-            overflow = TextOverflow.Ellipsis,
-            modifier = Modifier.padding(start = 12.dp)
-        )
-    }
-}
-
-private fun openApkWithChooser(context: android.content.Context, apkFile: java.io.File) {
-    try {
-        val uri = androidx.core.content.FileProvider.getUriForFile(
-            context, "${context.packageName}.fileprovider", apkFile
-        )
-        val sendIntent = android.content.Intent(android.content.Intent.ACTION_SEND).apply {
-
-            type = "*/*"
-            putExtra(android.content.Intent.EXTRA_STREAM, uri)
-
-            clipData = android.content.ClipData.newRawUri(apkFile.name, uri)
-            addFlags(android.content.Intent.FLAG_GRANT_READ_URI_PERMISSION)
-        }
-        val chooser = android.content.Intent.createChooser(sendIntent, Strings.buildSummaryOpenWith)
-        chooser.addFlags(android.content.Intent.FLAG_ACTIVITY_NEW_TASK)
-        context.startActivity(chooser)
-    } catch (e: Exception) {
-        AppLogger.e("BuildSummaryCard", "openApkWithChooser failed", e)
-        android.widget.Toast.makeText(
-            context, e.localizedMessage ?: "Open failed",
-            android.widget.Toast.LENGTH_SHORT
-        ).show()
-    }
-}
-
-@Composable
-fun EngineSelectionCard(
-    selectedEngine: String,
-    isGeckoDownloaded: Boolean,
-    onEngineSelected: (String) -> Unit
-) {
-    Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-        Text(
-            Strings.engineSelectTitle,
-            style = MaterialTheme.typography.titleSmall
-        )
-        Text(
-            Strings.engineSelectDesc,
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .clip(RoundedCornerShape(WtaRadius.Control))
-                .clickable { onEngineSelected("SYSTEM_WEBVIEW") }
-                .padding(8.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            RadioButton(
-                selected = selectedEngine == "SYSTEM_WEBVIEW",
-                onClick = { onEngineSelected("SYSTEM_WEBVIEW") }
-            )
-            Spacer(Modifier.width(8.dp))
-            Column(Modifier.weight(weight = 1f, fill = true)) {
-                Text(Strings.engineSystemWebView, style = MaterialTheme.typography.bodyMedium)
-                Text(
-                    Strings.engineSystemWebViewDesc,
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-        }
-
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .clip(RoundedCornerShape(WtaRadius.Control))
-                .clickable {
-                    if (isGeckoDownloaded) onEngineSelected("GECKOVIEW")
-                }
-                .padding(8.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            RadioButton(
-                selected = selectedEngine == "GECKOVIEW",
-                onClick = { if (isGeckoDownloaded) onEngineSelected("GECKOVIEW") },
-                enabled = isGeckoDownloaded
-            )
-            Spacer(Modifier.width(8.dp))
-            Column(Modifier.weight(weight = 1f, fill = true)) {
-                Row(verticalAlignment = Alignment.CenterVertically) {
-                    Text(
-                        Strings.engineGeckoView,
-                        style = MaterialTheme.typography.bodyMedium,
-                        color = if (isGeckoDownloaded) MaterialTheme.colorScheme.onSurface
-                                else MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                    if (isGeckoDownloaded) {
-                        Spacer(Modifier.width(6.dp))
-                        WtaBadge(
-                            text = Strings.engineReady,
-                            containerColor = MaterialTheme.colorScheme.primaryContainer,
-                            contentColor = MaterialTheme.colorScheme.onPrimaryContainer
-                        )
-                    }
-                }
-                if (!isGeckoDownloaded) {
-                    Text(
-                        Strings.engineGeckoNotDownloaded,
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.error
-                    )
-                } else {
-                    Text(
-                        Strings.engineApkSizeWarning.replace("%s", com.webtoapp.core.engine.EngineType.GECKOVIEW.estimatedSizeMb.toString()),
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                }
-            }
-        }
-    }
-}
